Refactor common portion of backend task.
