+++ infra/index.ts
@@ -5,7 +5,8 @@ import {createRepos} from "./ecr";
 import {createDatabase} from "./database";
 import {createCluster} from "./cluster";
 import {createAlb} from "./alb";
-import {createBackendTaskConfig, createBackend} from "./backend";
+import {createBackendTaskConfig} from "./backendtask";
+import {createBackend} from "./backend";
 import {createFrontend} from "./frontend";
 import {createMonitoring} from "./monitoring";
 import {createGraphCheck} from "./graphcheck";
